fix: rename CJS scripts to .cjs in ESM packages
Packages with "type": "module" treat .js files as ESM, causing
ReferenceError for require/module.exports. Rename affected scripts
and configs to .cjs and update all references.
