+++ .github/workflows/pipeline.yml
@@ -149,7 +149,7 @@ jobs:
                   git status
             - name: Sync MCP manifest versions
               if: steps.changesetVersion.outputs.changes == 'true'
-              run: node packages/fiori-mcp-server/scripts/sync-mcp-manifests.js
+              run: node packages/fiori-mcp-server/scripts/sync-mcp-manifests.cjs
             - name: Commit and push changes
               if: steps.changesetVersion.outputs.changes == 'true'
               run: |
