+++ packages/fiori-mcp-server/scripts/sync-mcp-manifests.cjs
