+++ packages/odata-annotation-core-types/lint-staged.config.cjs
