+++ packages/odata-annotation-core/lint-staged.config.cjs
