+++ packages/odata-entity-model/lint-staged.config.cjs
